--- jbugs-business/src/main/java/edu/msg/ro/service/UserManagement.java
@@ -1,51 +1,0 @@
-package edu.msg.ro.service;
-
-import edu.msg.ro.exceptions.BusinessException;
-import edu.msg.ro.transfer.UserDTO;
-
-import java.util.List;
-
-public interface UserManagement {
-
-    /**
-     * Method is used for persisting an user from an userDTO.
-     * It generates the username and does the validations.
-     * @param userDTO user information
-     * @return the newly created entity as a userDTO
-     */
-    UserDTO createUser(UserDTO userDTO) throws BusinessException;
-
-    /**
-     * Deactivates a user, restricting the access of said user to the app.
-     * @param username
-     */
-    void deactivateUser(String username);
-
-    /**
-     * Activates a deactivated user.
-     * @param username
-     */
-    void activateUser(String username);
-
-
-    /**
-     * @return a list of DTOs containing information about users.
-     */
-    List<UserDTO> getAllUsers();
-
-
-    /**
-     * Tries to log in a user.
-     * @param username
-     * @param password
-     * @return UserDTO of said user
-     * @throws BusinessException in case the user is not found or the password is wrong.
-     */
-    UserDTO login(String username, String password) throws BusinessException;
-
-
-
-
-
-
-}
